+++ packages/app/src/report/components/ReportPage.tsx
@@ -1,12 +1,12 @@
 import { FC, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { AnalysisErrorDialog } from "../../common/AnalysisErrorDialog";
-import { CodeTotalDrawer } from "../../common/CodeTotalDrawer";
-import { NewAnalysisDialog } from "../../common/NewAnalysisDialog";
+import { clearReport } from "../actions/clear-report-actions";
 import { initReport } from "../actions/init-report-action";
-import { LinterInfo } from "./drawer/LinterInfo";
-import { ReportHeader } from "./header/ReportHeader";
+import { LinterDrawer } from "./drawer/LinterDrawer";
 import { CodeDialog } from "./header/CodeDialog";
+import { NewAnalysisDialog } from "./header/NewAnalysisDialog";
+import { ReportHeader } from "./header/ReportHeader";
 import { ReportTabs } from "./tabs/ReportTabs";
 
 const ReportPage: FC = () => {
@@ -25,13 +25,15 @@ const ReportPage: FC = () => {
     })();
   }, [requestId, navigate]);
 
+  useEffect(() => {
+    return clearReport;
+  });
+
   return (
     <div style={{ paddingBlockEnd: 60 }}>
       <ReportHeader ready />
       <ReportTabs />
-      <CodeTotalDrawer>
-        <LinterInfo />
-      </CodeTotalDrawer>
+      <LinterDrawer />
       <AnalysisErrorDialog />
       <NewAnalysisDialog />
       <CodeDialog />